requirements.txt
groq>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
websockets>=13.0
pydantic>=2.9.0
httpx>=0.27.0
jinja2>=3.1.4
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
